+++ laxar-vue-adapter.js
@@ -13,7 +13,28 @@
 
 import Vue from 'vue';
 
-const WIDGET_PROPERTY = '_widget';
+const WIDGET_PROPERTY = '_axWidget';
+const INJECTIONS_PROPERTY = '_axWidgetInjections';
+
+/**
+ * Produces a Vue.js mixin that allows widgets to declare and access their injections.
+ *
+ * @param {...String} injections
+ *    a list of injection names. The respective services will be available in the same
+ *    order to the instance as `this.$injections`.
+ *
+ * @return {Object}
+ *    a Vue.js mixin that allowws the adapter to register the specified injections
+ */
+export function injections( ...injections ) {
+   return {
+      beforeCreate() {
+         this[ INJECTIONS_PROPERTY ] = injections;
+      }
+   };
+}
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 export const technology = 'vue';
 
@@ -43,9 +64,18 @@ export function bootstrap( _, adapterServices ) {
       const provider = artifactProvider.forWidget( widgetName );
       const widget = services.axContext.widget;
       const mixins = [
-         { beforeCreate() { this[ WIDGET_PROPERTY ] = widget; } },
+         {
+            beforeCreate() {
+               this[ WIDGET_PROPERTY ] = widget;
+            }
+         },
          widgetInjectionsMixin,
-         { beforeCreate() { provideServices( /* this.$options.injections */ services ); } }
+         {
+            beforeCreate() {
+               services.vueComponent = this;
+               provideServices( services );
+            }
+         }
       ];
 
       widgetServices[ widget.id ] = services;
@@ -173,17 +203,12 @@ export function bootstrap( _, adapterServices ) {
 
 function injectionsMixin( serviceFactory ) {
    return {
+      // run right before component creation to make sure that `injections` are available
       beforeCreate() {
-         const injections = this.$options.injections || [];
-
-         this.$injections = [];
-         this.$options.injections = {};
-
-         injections.forEach( injection => {
-            const service = serviceFactory.call( this, injection );
-            this.$injections.push( service );
-            this.$options.injections[ injection ] = service;
-         } );
+         const injections = this[ INJECTIONS_PROPERTY ];
+         if( injections ) {
+            this.$injections = injections.map( name => serviceFactory.call( this, name ) );
+         }
       }
    };
 }