+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ## Last Changes
+
+- [#1](https://github.com/LaxarJS/laxar-vue-adapter/issues/1): initial Vue.js technology adapter